Use the sass-bootstrap bower component
This requires a hack to include the component's sass directory to be
added to the sass load path. It was a simple line in application.rb, but
is not the most elegant solution.

I also considered adding `vendor/assets/components` to the load path then
including it in the scss file as `@import "sass-bootstrap/lib/bootstrap";`
but I am not sure which is cleaner.